refactor: cli package into analysis-utils and cli
