+++ packages/analysis-utils/.eslintrc.json
@@ -1,0 +1,7 @@
+{
+  "rules": {
+    "no-console": "off",
+    "no-await-in-loop": "off"
+  },
+  "ignorePatterns": ["dist/**", "dist-types/**"]
+}
