+++ packages/analysis-utils/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "@ps-analysis-tool/analysis-utils",
+  "version": "0.8.0",
+  "description": "A package for common utilities that are being used in multiple packages",
+  "main": "./dist/index.js",
+  "types": "./dist-types/index.d.ts",
+  "source": "./src/index.ts",
+  "customExports": {
+    ".": {
+      "default": "./src/index.ts"
+    }
+  },
+  "scripts": {
+    "build": "tsc",
+    "dev": "tsc-watch"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/GoogleChromeLabs/ps-analysis-tool"
+  },
+  "author": "",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/GoogleChromeLabs/ps-analysis-tool/issues"
+  },
+  "homepage": "https://github.com/GoogleChromeLabs/ps-analysis-tool",
+  "dependencies": {
+    "@ps-analysis-tool/common": "*",
+    "wappalyzer": "^6.10.66"
+  },
+  "devDependencies": {
+  }
+}
