+++ packages/analysis-utils/src/index.ts
@@ -14,27 +14,4 @@
  * limitations under the License.
  */
 
-/**
- * External dependencies.
- */
-import fs from 'fs/promises';
-import path from 'path';
-import { CookieDatabase } from '@ps-analysis-tool/common';
-
-/**
- * Internal dependencies.
- */
-
-/**
- * Fetch dictionary from local data folder.
- * @returns {Promise<CookieDatabase>} Open Cookie Data base
- */
-export async function fetchDictionary(): Promise<CookieDatabase> {
-  const data = JSON.parse(
-    await fs.readFile(path.resolve('./assets/data/open-cookie-database.json'), {
-      encoding: 'utf8',
-    })
-  );
-
-  return data;
-}
+export * from './procedures';
