+++ packages/analysis-utils/tsconfig.json
@@ -2,16 +2,15 @@
   "compilerOptions": {
     "rootDir": "src",
     "target": "es6",
+    "lib": ["es2021"],
     "module": "commonjs",
-    "outDir": "../../dist/cli",
+    "outDir": "dist",
+    "declarationDir": "dist-types",
+    "composite": true,
     "strict": true,
     "sourceMap": true,
     "esModuleInterop": true,
     "moduleResolution": "node"
   },
-  "references": [
-    { "path": "../common" },
-    { "path": "../i18n" },
-    { "path": "../analysis-utils" }
-  ]
+  "exclude": ["**/tests/**/*.ts", "dist/**", "dist-types/**"]
 }
