+++ packages/cli/src/index.ts
@@ -23,22 +23,24 @@ import { ensureFile, writeFile } from 'fs-extra';
 import Spinnies from 'spinnies';
 import { spawn } from 'child_process';
 import path from 'path';
-import { CompleteJson } from '@ps-analysis-tool/common';
+import { CompleteJson, delay } from '@ps-analysis-tool/common';
+import {
+  analyzeCookiesUrlsInBatches,
+  analyzeTechnologiesUrlsInBatches,
+} from '@ps-analysis-tool/analysis-utils';
 
 /**
  * Internal dependencies.
  */
-import Utility from './utils/utility';
-import { analyzeCookiesUrlsInBatches } from './procedures/analyzeCookieUrlsInBatches';
-import { analyzeTechnologiesUrlsInBatches } from './procedures/analyzeTechnologiesUrlsInBatches';
 import {
   fetchDictionary,
-  delay,
   getUrlListFromArgs,
   validateArgs,
   saveCSVReports,
+  askUserInput,
+  checkPortInUse,
+  generatePrefix,
 } from './utils';
-import { checkPortInUse } from './utils/checkPortInUse';
 
 events.EventEmitter.defaultMaxListeners = 15;
 
@@ -138,7 +140,7 @@ const startDashboardServer = async (dir: string, port: number) => {
 
   const prefix =
     url || sitemapUrl
-      ? Utility.generatePrefix(url || sitemapUrl)
+      ? generatePrefix(url || sitemapUrl)
       : path.parse(csvPath || sitemapPath).name;
 
   let outputDir;
@@ -168,7 +170,7 @@ const startDashboardServer = async (dir: string, port: number) => {
     let userInput: string | null = null;
 
     if (!shouldSkipPrompts && !numberOfUrlsInput) {
-      userInput = await Utility.askUserInput(
+      userInput = await askUserInput(
         `Provided ${sitemapUrl || sitemapPath ? 'Sitemap' : 'CSV file'} has ${
           urls.length
         } pages. Please enter the number of pages you want to analyze (Default ${
@@ -179,7 +181,7 @@ const startDashboardServer = async (dir: string, port: number) => {
       numberOfUrls =
         userInput && isNaN(parseInt(userInput))
           ? urls.length
-          : parseInt(userInput);
+          : parseInt(userInput as string);
     } else if (numberOfUrlsInput) {
       console.log(`Analysing ${numberOfUrlsInput} urls.`);
       numberOfUrls = parseInt(numberOfUrlsInput);
