+++ packages/cli/src/utils/getUrlListFromArgs.ts
@@ -18,10 +18,10 @@
  */
 import { readFile } from 'fs-extra';
 import { parseStringPromise } from 'xml2js';
+import getUrlsFromSitemap from './getUrlsfromSitemap';
 /**
  * Internal dependencies.
  */
-import Utility from './utility';
 
 const parseUrlsFromSitemap = async (sitemapUrl: string, spinnies: any) => {
   spinnies?.add('sitemap-spinner', {
@@ -29,7 +29,7 @@ const parseUrlsFromSitemap = async (sitemapUrl: string, spinnies: any) => {
   });
 
   try {
-    const _urls = await Utility.getUrlsFromSitemap(sitemapUrl);
+    const _urls = await getUrlsFromSitemap(sitemapUrl);
     spinnies?.succeed('sitemap-spinner', {
       text: 'Done parsing Sitemap',
     });
