+++ packages/cli/src/utils/index.ts
@@ -13,9 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { default as delay } from './delay';
 export { fetchDictionary } from './fetchCookieDictionary';
 export { default as validateArgs } from './validateArgs';
 export { default as getUrlListFromArgs } from './getUrlListFromArgs';
 export { default as generateCSVFiles } from './generateCSVfiles';
 export { default as saveCSVReports } from './saveCSVReports';
+export { default as getUrlsFromSitemap } from './getUrlListFromArgs';
+export { default as askUserInput } from './askUserInput';
+export { checkPortInUse } from './checkPortInUse';
+export { default as generatePrefix } from './generatePrefix';
