+++ packages/common/src/index.ts
@@ -37,6 +37,7 @@ export { default as executeTaskInDevToolWorker } from './worker/executeTaskInDev
 export { default as getValueByKey } from './utils/getValueByKey';
 export * from './utils/contextSelector';
 export { default as addUTMParams } from './utils/addUTMParams';
+export { default as delay } from './utils/delay';
 export * from './worker/enums';
 export * from './utils/generateReports';
 export * from './cookies.types';
